moved credential to cfme
